chore: migrate `extra_flags`
