+++ crates/moon/tests/test_cases/extra_flags/lib/hello.mbt
