+++ crates/moon/tests/test_cases/extra_flags/lib/moon.pkg.json
