+++ crates/moon/tests/test_cases/extra_flags/main/main.mbt
