+++ crates/moon/tests/test_cases/extra_flags/main/moon.pkg.json
